+++ src/utils/formatters.js
@@ -1,9 +1,11 @@
 import path from "path"
 import { getConfigFile, humanizeAmount } from "medusa-core-utils"
 
-const { configModule } = getConfigFile(path.resolve("."), "medusa-config")
-const { plugins } = configModule
-const { options } = plugins.find(e => e.resolve === 'medusa-fulfillment-shippo')
+const {
+  configModule: { projectConfig, plugins },
+} = getConfigFile(path.resolve("."), "medusa-config")
+const config = plugins.find((e) => e.resolve === "medusa-fulfillment-shippo")
+const options = config?.options ?? projectConfig
 
 /** Makes a flat product object from a LineItem
  * @param {LineItem} - LineItem object
@@ -62,11 +64,10 @@ export const shippoAddress = (address, email) => ({
   validate: address.country_code == "us" ?? true,
 })
 
-export const shippoOrder = async (order, lineItems, parcel) => {
+export const shippoOrder = (order, lineItems, parcel) => {
   const toAddress = shippoAddress(order.shipping_address, order.email)
   const currencyCode = order.currency_code.toUpperCase()
   const shippingOptionName = order.shipping_methods[0].shipping_option.name
-
   const totalWeight = lineItems
     .map((e) => e.weight * e.quantity)
     .reduce((sum, current) => sum + current, 0)